pyproject.toml
[project]
name = "nexvec"
version = "2.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "google-genai>=1.68.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "pdfplumber>=0.11",
    "pgvector>=0.3",
    "psycopg2-binary>=2.9",
    "pydantic>=2.6",
    "pymysql>=1.1",
    "pytest>=8.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7.4",
    "uvicorn[standard]>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
